use accesscontrol

+++ src/controllers/employeesController.js
@@ -1,4 +1,7 @@
+import jwt from 'jsonwebtoken';
+
 import db from '../models';
+import accesscontrol from '../policy/role';
 
 const User = db.users
 const Employee = db.employees
@@ -6,9 +9,19 @@ const Customer = db.customers
 
 // 1. get all employees
 const getAllEmployees = async (req, res) => {
+    const username = req.locals.username
     try {
-        let employees = await Employee.findAll({});
-        res.status(200).send(employees);    
+        //  get jobTitle
+        const jobTitle = getJobTitle(username);
+
+        // Check role permissions
+        const permission = accesscontrol.can(jobTitle).readAny('employees');
+        if(permission.granted) {
+            const employees = await Employee.findAll({});
+            res.status(200).send(employees);    
+        } else {
+            res.status(403).end();
+        }
     } catch (error) {
         console.log(error);
     }
@@ -17,12 +30,25 @@ const getAllEmployees = async (req, res) => {
 // 2. create a employee
 const createEmployee = async (req, res) => {
     const data = req.body;
+    const username = req.locals.username
     try {
-        const employee = await Employee.create(data);
-        res.status(201).json({
-            success: true,
-            data: employee
-        })
+        //  get jobTitle
+        const jobTitle = getJobTitle(username);
+
+        // Check role permissions
+        const permission = accesscontrol.can(jobTitle).createAny('employees');
+        if(permission.granted) {
+            const employee = await Employee.create(data);
+            const jwt_token = req.cookies.jwt_token;
+            const username = jwt.verify(jwt_token, process.env.ACCESS_TOKEN_SECRET); 
+            await User.update({ employeeNumber: employee.employeeNumber }, { where: { username: username }});
+            res.status(201).json({
+                success: true,
+                data: employee
+            })
+        } else {
+            res.status(403).end();
+        }
     } catch (error) {
         console.log(error);
     }
@@ -30,32 +56,42 @@ const createEmployee = async (req, res) => {
 
 // 3. delete all employees
 const deleteAllEmployees = async (req, res) => {
+    
     const CONSTRAINT_CUSTOMERS_EMPLOYEES = 'customers_ibfk_1';
     const CUSTOMERS_TABLE = 'customers';
     const EMPLOYEES_TABLE = 'employees';
     const EMPLOYEES_PK = 'employeeNumber';
     const CUSTOMERS_REF = 'salesRepEmployeeNumber';
     const VALUE_NULL_CUSTOMER = { salesRepEmployeeNumber: null };
-
+    const username = req.locals.username
     try {
-        await Customer.update(VALUE_NULL_CUSTOMER, { where: {}});
-        await db.sequelize.queryInterface.removeConstraint(CUSTOMERS_TABLE, CONSTRAINT_CUSTOMERS_EMPLOYEES);
-        await Employee.truncate();
-        await db.sequelize.queryInterface.addConstraint(CUSTOMERS_TABLE, {
-            fields: [CUSTOMERS_REF],
-            type: 'foreign key',
-            name: CONSTRAINT_CUSTOMERS_EMPLOYEES,
-            references: {
-                table: EMPLOYEES_TABLE,
-                field: EMPLOYEES_PK
-            },
-            onDelete: 'cascade',
-            onUpdate: 'cascade'
-        })
-        res.status(200).json({
-            success: true,
-            message: "Delete all employees"
-        });
+        //  get jobTitle
+        const jobTitle = getJobTitle(username);
+
+        // Check role permissions
+        const permission = accesscontrol.can(jobTitle).deleteAny('employees');
+        if(permission.granted) {
+            await Customer.update(VALUE_NULL_CUSTOMER, { where: {}});
+            await db.sequelize.queryInterface.removeConstraint(CUSTOMERS_TABLE, CONSTRAINT_CUSTOMERS_EMPLOYEES);
+            await Employee.truncate();
+            await db.sequelize.queryInterface.addConstraint(CUSTOMERS_TABLE, {
+                fields: [CUSTOMERS_REF],
+                type: 'foreign key',
+                name: CONSTRAINT_CUSTOMERS_EMPLOYEES,
+                references: {
+                    table: EMPLOYEES_TABLE,
+                    field: EMPLOYEES_PK
+                },
+                onDelete: 'cascade',
+                onUpdate: 'cascade'
+            })
+            res.status(200).json({
+                success: true,
+                message: "Delete all employees"
+            });
+        } else {
+            res.status(403).end();
+        }
     } catch (error) {
         console.log(error);
     }
@@ -64,9 +100,19 @@ const deleteAllEmployees = async (req, res) => {
 // 4. get one employee by Id
 const getOneEmployee = async (req, res) => {
     const employeeNumber = req.params.employeeNumber;
+    const username = req.locals.username
     try {
-        const employee = await Employee.findByPk(employeeNumber);
-        res.status(200).send(employee);
+        //  get jobTitle
+        const jobTitle = getJobTitle(username);
+
+        // Check role permissions
+        const permission = accesscontrol.can(jobTitle).readAny('employees');
+        if(permission.granted) {
+            const employee = await Employee.findByPk(employeeNumber);
+            res.status(200).send(employee);  
+        } else {
+            res.status(403).end();
+        }
     } catch (error) {
         console.log(error);
     }
@@ -76,18 +122,28 @@ const getOneEmployee = async (req, res) => {
 const updateOrCreateEmployee = async (req, res) => {
     const employeeNumber = req.params.employeeNumber;
     const data = req.body;
+    const username = req.locals.username
     try {
-        let employee = Employee.findByPk(employeeNumber);
-        if(!employee) {
-            employee = await Employee.create(data);
+        //  get jobTitle
+        const jobTitle = getJobTitle(username);
+
+        // Check role permissions
+        const permission = accesscontrol.can(jobTitle).updateAny('employees');
+        if(permission.granted) {
+            let employee = Employee.findByPk(employeeNumber);
+            if(!employee) {
+                employee = await Employee.create(data);
+            } else {
+                employee = await Employee.update(data, { where: { employeeNumber: employeeNumber }});
+            }
+            res.status(200).json({
+                success: true,
+                message: "Update or Create a employee",
+                data: employee
+            });
         } else {
-            employee = await Employee.update(data, { where: { employeeNumber: employeeNumber }});
+            res.status(403).end();
         }
-        res.status(200).json({
-            success: true,
-            message: "Update or Create a employee",
-            data: employee
-        });
     } catch (error) {
         console.log(error);
     }
@@ -97,13 +153,25 @@ const updateOrCreateEmployee = async (req, res) => {
 const updateEmployee = async (req, res) => {
     const employeeNumber = req.params.employeeNumber;
     const data = req.body;
+    const username = req.locals.username
+
     try {
-        const employee = await Employee.update(data, { where: { employeeNumber: employeeNumber }});
-        res.status(200).json({
-            success: true,
-            message: "Update a employee",
-            data: employee
-        })
+
+        //  get jobTitle
+        const jobTitle = getJobTitle(username);
+
+        // Check role permissions
+        const permission = accesscontrol.can(jobTitle).updateAny('employees');
+        if(permission.granted) {
+            const employee = await Employee.update(data, { where: { employeeNumber: employeeNumber }});
+            res.status(200).json({
+                success: true,
+                message: "Update a employee",
+                data: employee
+            })  
+        } else {
+            res.status(403).end();
+        }        
     } catch (error) {
         console.log(error);
     }
@@ -114,14 +182,25 @@ const deleteOneEmployee = async (req, res) => {
     const employeeNumber = req.params.employeeNumber;
     const VALUE_NULL_CUSTOMER = { salesRepEmployeeNumber: null };
     const VALUE_NULL_EMPLOYEE = { reportsTo: null };
+    const username = req.locals.username
     try {
-        await Employee.update(VALUE_NULL_EMPLOYEE, { where: { reportsTo: employeeNumber }});
-        await Customer.update(VALUE_NULL_CUSTOMER, { where: { salesRepEmployeeNumber: employeeNumber }});
-        await Employee.destroy({ where: { employeeNumber: employeeNumber}});
-        res.status(200).json({
-            success: true,
-            message: "Delete a employee"
-        });
+
+        //  get jobTitle
+        const jobTitle = getJobTitle(username);
+
+        // Check role permissions
+        const permission = accesscontrol.can(jobTitle).deleteAny('employees');
+        if(permission.granted) {
+            await Employee.update(VALUE_NULL_EMPLOYEE, { where: { reportsTo: employeeNumber }});
+            await Customer.update(VALUE_NULL_CUSTOMER, { where: { salesRepEmployeeNumber: employeeNumber }});
+            await Employee.destroy({ where: { employeeNumber: employeeNumber}});
+            res.status(200).json({
+                success: true,
+                message: "Delete a employee"
+            });
+        } else {
+            res.status(403).end();
+        }
     } catch (error) {
         console.log(error);
     }
@@ -130,7 +209,20 @@ const deleteOneEmployee = async (req, res) => {
 // 8. get all customers of an employee
 const getCusomtersOfEmployee = async (req, res) => {
     const employeeNumber = req.params.employeeNumber;
+    const username = req.locals.username;
+
     try {
+        //  get jobTitle
+        const jobTitle = getJobTitle(username);
+
+        // Check role permissions
+        const permission = accesscontrol.can(jobTitle).readOwn('employees');
+        if(permission.granted) {
+            const employees = await Employee.findAll({});
+            res.status(200).send(employees);    
+        } else {
+            res.status(403).end();
+        }        
         const data = await Employee.findAll({
             include: [{
                 model: Customer,
@@ -144,7 +236,7 @@ const getCusomtersOfEmployee = async (req, res) => {
     }
 }
 
-// 9. get all staff of an employee
+// 9. BONUS get all staff of an employee by reportsTo
 const getStaffsOfEmployee = async (req, res) => {
     const employeeNumber = req.params.employeeNumber;
     try {
@@ -159,6 +251,15 @@ const getStaffsOfEmployee = async (req, res) => {
     } catch (error) {
         console.log(error);
     }
+}
+
+
+
+// get jobTitle
+async function getJobTitle(username) {
+    const user = await User.findByPk(username);
+    const employee = await Employee.findByPk(user.employeeNumber);
+    return employee.jobTitle;
 }
 
 module.exports = {
